+++ ios/RNIapIos.m
@@ -109,7 +109,7 @@ RCT_EXPORT_MODULE();
 
 - (NSArray<NSString *> *)supportedEvents
 {
-    return @[@"iap-purchase-event", @"iap-promoted-product", @"purchase-updated", @"purchase-error"];
+    return @[@"iap-promoted-product", @"purchase-updated", @"purchase-error"];
 }
 
 RCT_EXPORT_METHOD(canMakePayments:(RCTPromiseResolveBlock)resolve
@@ -150,7 +150,6 @@ RCT_EXPORT_METHOD(buyProduct:(NSString*)sku
     if (product) {
         SKMutablePayment *payment = [SKMutablePayment paymentWithProduct:product];
         [[SKPaymentQueue defaultQueue] addPayment:payment];
-        [self addPromiseForKey:RCTKeyForInstance(payment.productIdentifier) resolve:resolve reject:reject];
     } else {
         if (hasListeners) {
             NSDictionary *err = [NSDictionary dictionaryWithObjectsAndKeys:
@@ -194,7 +193,6 @@ RCT_EXPORT_METHOD(buyProductWithOffer:(NSString*)sku
         #endif
         payment.applicationUsername = usernameHash;
         [[SKPaymentQueue defaultQueue] addPayment:payment];
-        [self addPromiseForKey:RCTKeyForInstance(payment.productIdentifier) resolve:resolve reject:reject];
     } else {
         if (hasListeners) {
             NSDictionary *err = [NSDictionary dictionaryWithObjectsAndKeys:
@@ -225,37 +223,6 @@ RCT_EXPORT_METHOD(buyProductWithQuantityIOS:(NSString*)sku
         SKMutablePayment *payment = [SKMutablePayment paymentWithProduct:product];
         payment.quantity = quantity;
         [[SKPaymentQueue defaultQueue] addPayment:payment];
-        [self addPromiseForKey:RCTKeyForInstance(payment.productIdentifier) resolve:resolve reject:reject];
-    } else {
-        if (hasListeners) {
-            NSDictionary *err = [NSDictionary dictionaryWithObjectsAndKeys:
-                                 @"Invalid product ID.", @"debugMessage",
-                                 @"Invalid product ID.", @"message",
-                                 @"E_DEVELOPER_ERROR", @"code",
-                                 nil
-                                 ];
-            [self sendEventWithName:@"purchase-error" body:err];
-        }
-        reject(@"E_DEVELOPER_ERROR", @"Invalid product ID.", nil);
-    }
-}
-
-// The following buyProductWithoutAutoConfirm seems to be completely unused and identical to buyProduct. It's existance is confusing so could we remove it?
-RCT_EXPORT_METHOD(buyProductWithoutAutoConfirm:(NSString*)sku
-                  resolve:(RCTPromiseResolveBlock)resolve
-                  reject:(RCTPromiseRejectBlock)reject) {
-    NSLog(@"\n\n\n  buyProductWithoutAutoConfirm  \n\n.");
-    SKProduct *product;
-    for (SKProduct *p in validProducts) {
-        if([sku isEqualToString:p.productIdentifier]) {
-            product = p;
-            break;
-        }
-    }
-    if (product) {
-        SKMutablePayment *payment = [SKMutablePayment paymentWithProduct:product];
-        [[SKPaymentQueue defaultQueue] addPayment:payment];
-        [self addPromiseForKey:RCTKeyForInstance(payment.productIdentifier) resolve:resolve reject:reject];
     } else {
         if (hasListeners) {
             NSDictionary *err = [NSDictionary dictionaryWithObjectsAndKeys:
@@ -296,7 +263,6 @@ RCT_EXPORT_METHOD(buyPromotedProduct:(RCTPromiseResolveBlock)resolve
     if (promotedPayment) {
         NSLog(@"\n\n\n  ***  buy promoted product. \n\n.");
         [[SKPaymentQueue defaultQueue] addPayment:promotedPayment];
-        [self addPromiseForKey:RCTKeyForInstance(promotedPayment.productIdentifier) resolve:resolve reject:reject];
     } else {
         reject(@"E_DEVELOPER_ERROR", @"Invalid product ID.", nil);
     }
@@ -347,7 +313,6 @@ RCT_EXPORT_METHOD(getPendingTransactions:(RCTPromiseResolveBlock)resolve
 #pragma mark ===== StoreKit Delegate
 
 -(void)productsRequest:(SKProductsRequest *)request didReceiveResponse:(SKProductsResponse *)response {
-    // validProducts = response.products; // do not set directly. use addProduct methods.
     for (SKProduct* prod in response.products) {
         [self addProduct:prod];
     }
@@ -405,7 +370,7 @@ RCT_EXPORT_METHOD(getPendingTransactions:(RCTPromiseResolveBlock)resolve
                 NSLog(@"\n\n\n\n\n Purchase Successful !! \n\n\n\n\n.");
                 [self purchaseProcess:transaction];
                 break;
-            case SKPaymentTransactionStateRestored: // 기존 구매한 아이템 복구..
+            case SKPaymentTransactionStateRestored:
                 NSLog(@"Restored ");
                 [[SKPaymentQueue defaultQueue] finishTransaction:transaction];
                 break;
@@ -481,7 +446,6 @@ RCT_EXPORT_METHOD(getPendingTransactions:(RCTPromiseResolveBlock)resolve
         
         // additionally send event
         if (hasListeners) {
-            [self sendEventWithName:@"iap-purchase-event" body: purchase];
             [self sendEventWithName:@"purchase-updated" body: purchase];
         }
     }];